requirements.txt
quart
quart_cors
numpy==2.0.1
pillow==10.4.0
Augmentor==0.2.12